pyproject.toml
[tool.poetry]
name = "python3-tool-agents"
version = "0.1.0"
description = ""
authors = ["sairakausar <sairakausar3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-core = "^0.3.10"
langchain-community = "^0.3.2"
langchain-google-genai = "^2.0.1"
uvicorn = "^0.31.1"
python-dotenv = "^1.0.1"
openai = "^1.51.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
